requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM and ML
openai>=1.3.5
transformers>=4.36.0
torch>=2.2.0
peft>=0.7.0  # For LoRA
accelerate>=0.25.0

# Graph database
neo4j>=5.14.0
py2neo>=2021.2.3

# Document processing
PyMuPDF>=1.23.8
pandas>=2.1.3
numpy>=1.25.2

# AWS
boto3>=1.29.7
aioboto3>=11.3.0

# API and async
aiofiles>=23.2.1
httpx>=0.25.2
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.1
structlog>=23.2.0
prometheus-client>=0.19.0